feat: 全局反连迁移至设置中心，删除原弹窗组件
